requirements.txt
playwright==1.40.0
watchdog==3.0.0
jinja2==3.1.2
click==8.1.7
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
